pyproject.toml
[project]
name = "x-analysis-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tweepy>=4.14.0",
    "python-dotenv>=1.0.0",
    "textblob>=0.17.1"
]
